Jack0307

+++ src/main/java/com/oldFoodMan/demo/service/ShoppingCartService.java
@@ -76,37 +76,31 @@ public class ShoppingCartService {
 		ShoppingCart a = shopDao.findByProductIdAndMemberId(product_id, member_id);
 		int Product_amount = a.getProductAmount();
 		
-//		//判斷使用者有無超過商品庫存數量  3/6測試看看
-//		Integer stock = product.getProduct_stock();
-//		if(Product_amount > stock) {
-//			return null;
-//		}else {
-//			ShoppingCart cart = new ShoppingCart();
-//			cart.setMemberId(member);
-//			cart.setProductId(product);
-//			cart.setProductAmount(Product_amount+1);
-//			shopDao.delete(a);
-//			shopDao.save(cart);
-//			
-//			ShoppingCart newCart = shopDao.findByProductIdAndMemberId(product_id, member_id);
-//			return newCart;
-//		}
-		
-		
-		
-		ShoppingCart cart = new ShoppingCart();
-		cart.setMemberId(member);
-		cart.setProductId(product);
-		cart.setProductAmount(Product_amount+1);
-		shopDao.delete(a);
-		shopDao.save(cart);
-		
-		ShoppingCart newCart = shopDao.findByProductIdAndMemberId(product_id, member_id);
-		return newCart;
-		
+		//判斷使用者有無超過商品庫存數量 
+		Integer stock = product.getProduct_stock();
+		if(Product_amount >= stock) {
+			ShoppingCart cart = new ShoppingCart();
+			cart.setMemberId(member);
+			cart.setProductId(product);
+			cart.setProductAmount(stock);
+			shopDao.save(cart);
+			
+			ShoppingCart newCart = shopDao.findByProductIdAndMemberId(product_id, member_id);
+			return newCart;
+		}else {
+			ShoppingCart cart = new ShoppingCart();
+			cart.setMemberId(member);
+			cart.setProductId(product);
+			cart.setProductAmount(Product_amount+1);
+			shopDao.delete(a);
+			shopDao.save(cart);
+			
+			ShoppingCart newCart = shopDao.findByProductIdAndMemberId(product_id, member_id);
+			return newCart;
+		}
 	}
 	
-	public void updateLowProductAmount(Product product, Member member) {
+	public ShoppingCart updateLowProductAmount(Product product, Member member) {
 		Integer product_id = product.getProduct_id();
 		Integer member_id = member.getId();
 		System.out.println("product_id : " + product_id);
@@ -114,25 +108,25 @@ public class ShoppingCartService {
 		ShoppingCart a = shopDao.findByProductIdAndMemberId(product_id, member_id);
 		int Product_amount = a.getProductAmount();
 		
-//		//判斷使用者點擊數量有無低於0   3/6測試看看
-//		Integer stock = product.getProduct_stock();
-//		if(Product_amount <= 0) {
-//			return null;
-//		}else {
-//			ShoppingCart cart = new ShoppingCart();
-//			cart.setMemberId(member);
-//			cart.setProductId(product);
-//			cart.setProductAmount(Product_amount-1);
-//			shopDao.delete(a);
-//			shopDao.save(cart);
-//		}
-		
-		ShoppingCart cart = new ShoppingCart();
-		cart.setMemberId(member);
-		cart.setProductId(product);
-		cart.setProductAmount(Product_amount-1);
-		shopDao.delete(a);
-		shopDao.save(cart);
+		//判斷使用者點擊數量有無低於0   
+		if(Product_amount <= 1) {
+			ShoppingCart cart = new ShoppingCart();
+			cart.setMemberId(member);
+			cart.setProductId(product);
+			cart.setProductAmount(1);
+			shopDao.delete(a);
+			shopDao.save(cart);
+			return cart;
+		}else {
+			ShoppingCart cart = new ShoppingCart();
+			cart.setMemberId(member);
+			cart.setProductId(product);
+			cart.setProductAmount(Product_amount-1);
+			shopDao.delete(a);
+			shopDao.save(cart);
+			return cart;
+		}
+	
 	}
 	
 	public ShoppingCart updateAutoProductAmount(Product product, Member member, Integer inputVal) {
@@ -141,7 +135,6 @@ public class ShoppingCartService {
 		ShoppingCart a = shopDao.findByProductIdAndMemberId(product_id, member_id);
 		shopDao.delete(a);
 		
-		//要直接幫使用者改最小或最大嗎???還是回歸未更動前的數量?
 		Integer stock = product.getProduct_stock();
 		if(inputVal > stock) {
 			ShoppingCart cart = new ShoppingCart();
@@ -165,14 +158,25 @@ public class ShoppingCartService {
 		return shopDao.findByProductIdAndMemberId(product_id, member_id);
 	}
 	
-//	//點擊按鈕刪除單筆購物車商品 3/6試試看
-//	public void deleteFromCart(Product product, Member member) {
-//		Integer member_id = (Integer)member.getId();
-//		Integer product_id = product.getProduct_id();
-//		
-//		ShoppingCart cart = shopDao.findByProductIdAndMemberId(product_id, member_id);
-//		shopDao.delete(cart);
-//	}
+	public void deleteFromCart(Product product, Member member) {
+		Integer member_id = (Integer)member.getId();
+		Integer product_id = product.getProduct_id();
+		
+		ShoppingCart cart = shopDao.findByProductIdAndMemberId(product_id, member_id);
+		shopDao.delete(cart);
+	}
+	
+	//此方法為後台刪除產品時，先刪除使用者購物車內同一產品，避免外鍵約束
+	public void deleteAllOneFromCart(Integer productId) {
+		System.out.println("TEST:  !!!!" + productId);
+		shopDao.deleteAllOneByProductId(productId);
+	}
+	
+	public ShoppingCart findMyOneProduct(Integer productId, Member member) {
+		Integer member_id = (Integer)member.getId();
+		ShoppingCart cart = shopDao.findByProductIdAndMemberId(productId, member_id);
+		return cart;
+	}
 	
 	
 }
